+++ src/main/java/com/example/ShopAppSelling/Repositories/ProductImageRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.ShopAppSelling.Models.ProductImage;
 
-public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
+public interface ProductImageRepository extends JpaRepository<ProductImage, Double> {
 
-    List<ProductImage> findByProductId(Long productId);
+    List<ProductImage> findByProductId(Double productId);
 }
