+++ src/main/java/com/example/ShopAppSelling/Services/CategoryService.java
@@ -25,7 +25,7 @@ public class CategoryService implements ICategoryService {
     }
 
     @Override
-    public Category getCategoryById(Long id) {
+    public Category getCategoryById(Double id) {
         return categoryRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Category not found with id: " + id));
     }
@@ -36,14 +36,14 @@ public class CategoryService implements ICategoryService {
     }
 
     @Override
-    public Category updateCategory(Long categoryId, CategoriesDTO categoriesDTO) {
+    public Category updateCategory(Double categoryId, CategoriesDTO categoriesDTO) {
         Category existingCategory = getCategoryById(categoryId);
         existingCategory.setName(categoriesDTO.getName());
         return categoryRepository.save(existingCategory);
     }
 
     @Override
-    public void deleteCategory(Long id) {
+    public void deleteCategory(Double id) {
         categoryRepository.deleteById(id);
     }
